Please Don't Make the Shield Move the Main Camera Next Time...
Seriously, don't do that.  It makes no sense why we were moving the
camera with the shield's script...
Updated the level block prefabs to conform to the following:
-15 units in width
-Center of the prefab is at 0, 0, 0
-"Ground Level" is considered to be Y = -2.5

Added a Scene to randomly generate a level (RandomGeneration.unity)
Enemies are now standing still rather than running in place

TODO:
Replace enemies with the newer ranged enemies (but keep the detections
zones about the way that they are now)
Fix the player collisions (The player collides and stops with EVERYTHING
but the enemies and their detection zones.  This means slow blocks, the
death block, and probably some other collisions aren't working properly
either)
Improve the camera (Right now it just follows the character's X
position, but it should move independently of the player so that we can
move the player around the screen when they slow down or dash)
Find some way to weight which blocks will spawn so that we can have
certain blocks spawning more often.
Add collectibles to the prefabs so that they will spawn too.
Make the player destroy bashable objects while they are shield bashing
(i.e. doors)