Procedurally Generate Prefabs with Different Lengths
Prefabs can now be randomly generated no matter how wide they are.
There are a few rules to this:
-Prefabs must be centered on their parent objects
-The parent needs to have a "PrefabLength.cs" script with the length of
the prefab set in the public length variable

Some things to consider:
-Make sure the spawner is far enough away from the camera so that large
prefabs won't suddenly appear on screen
Make sure prefabs are large enough that they won't be skipped if the
camera is moving fast enough.

Also in this:
-Tap on the left 30% of the screen to jump, swipe anywhere to shield