Added Different Lengths of the FlatGround prefab to Test Spawning
There are now different lengths of the FlatGround prefab if you want to
test the spawning.  I believe I have it spawning in blocks seamlessly
(no gaps or overlaps), so as long as we list the EXACT length of the
prefab in their PrefabLength script we should be good.

Also, I set it to always spawn a FlatGround15m prefab first just because
it fits with the ground built into the level and where the spawner
starts