Huge Update to Camera
Added a "ghost" object that now moves at the speed of the game. This
will control how fast the screen should be scrolling at any given
moment.
The camera follows the ghost instead of the player, so if the player
stops moving, they'll move beyond the screen boundaries and die.
Added a new "Ghost" tag for the ghost object
Added detection zones to the downwards transitional blocks  These will
move the ghost object downwards along a slope while it is inside the
detection area.  Once it leaves it gets reset to it's current movement
speed and forward direction.
Modified the "PrefabLengths" script to now be "PrefabAttributes" because
it can now contain any other information about the prefabs that we need
(i.e. Height, Transition type, etc)
Made an SFX and BGM folder in Audio to separate out what we have for
each.
Improved the SpawnScript to make sure that prefabs are spawned
seamlessly while the spawner is moving down slopes with the camera.
Separated the Destroyer script into "Player" and "Level" Destroyer
scripts that affect their respective game objects (except the level
destroyer can also "kill" the player).