Finished Endless Tutorial
Added some logic to the spawner script to start randomly generating
terrain if the level loaded isn't the tutorial level or if the spawner's
x position is > 350 (hard coded value.  This is is how big the tutorial
level is horizontally)
Added a few new prefabs that are variations of the enemy below:
PassThrough/LongPassThrough - Do not require the user to jump, but the
enemy can shoot through the platform.
ShortGap - Shorter jump and much shorter detection area for the enemy,
but the alert area is the same as the others so it still feels fair
(right?)

Edited some small values in other prefabs (like the detection areas of
some ranged enemies) to be nicer looking numbers.